driver-service/requirements.txt
mongoengine==0.27.0
pymongo==3.12.3
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3 